feat: 翻譯進度改為及時統計

+++ src/main/java/com/pino/translateitapi/model/dto/Project.java
@@ -21,7 +21,7 @@ public class Project {
     /**
      * 進度(%)
      */
-    private Integer progressRate;
+    private long progressRate;
     /**
      * 建立日期
      */
